+++ HotaMechanics/constants.h
@@ -7,17 +7,17 @@ namespace HotaMechanics::Constants {
 
 	static const int16_t MANA_PER_KGD = 10;
 	static const int16_t MAX_ARMY_SIZE = 7;
-	static const int16_t FIELD_COLS = 17;
-	static const int16_t FIELD_ROWS = 11;
-	static constexpr int16_t FIELD_SIZE = FIELD_COLS * FIELD_ROWS;
-	static const int16_t INVALID_HEX_ID = FIELD_SIZE;
+	static const uint8_t FIELD_COLS = 17;
+	static const uint8_t FIELD_ROWS = 11;
+	static constexpr uint8_t FIELD_SIZE = FIELD_COLS * FIELD_ROWS;
+	static const uint8_t INVALID_HEX_ID = FIELD_SIZE;
 	static const int16_t GUID_OFFSET = 21;
 
-	static std::vector<int16_t> EMPTY_PATH;
-	static std::vector<int16_t> EMPTY_VEC;
+	static std::vector<uint8_t> EMPTY_PATH;
+	static std::vector<uint8_t> EMPTY_VEC;
 
-	using AdjacentArray = std::array<int16_t, 6>;
-	using FieldArray = std::array<int, FIELD_SIZE + 1>;
+	using AdjacentArray = std::array<uint8_t, 6>;
+	using FieldArray = std::array<uint8_t, FIELD_SIZE + 1>;
 	using FieldFlagArray = std::array<bool, FIELD_SIZE + 1>;
 
 	enum class AIDifficulty : int8_t {
